requirements.txt
# Basic libraries
streamlit==1.22.0

# LangChain and dependencies
langchain==0.0.21
langchain-core==0.0.13  # Ensure a compatible version of langchain-core
langchain-community==0.0.21

# Use the latest compatible version of langchain-huggingface
langchain-huggingface==0.0.3

# HuggingFace libraries
transformers==4.30.2
datasets==2.12.0

# FAISS (Facebook AI Similarity Search)
faiss-cpu==1.7.3

# Chroma (for document embedding search)
chromadb==0.3.23

# Pinecone (for vector databases)
pinecone-client==2.2.1

# NLP libraries
nltk==3.8.1
spacy==3.5.2
scikit-learn==1.2.2

# Deep learning libraries
torch==2.0.1
tensorflow==2.12.0

# Streamlit for the web app
streamlit==1.22.0

# Database interaction
sqlalchemy==2.0.12
psycopg2-binary==2.9.6  # PostgreSQL database adapter for Python

# API interaction
requests==2.28.2
fastapi==0.96.0

# Other necessary libraries
numpy==1.24.2
pandas==1.5.3

# Utility libraries
python-dotenv==1.0.0
pydantic==1.10.7

# Additional useful libraries
sentence-transformers==2.2.2  # For sentence embeddings
tqdm==4.65.0  # Progress bar
click==8.1.3  # Command line utility
matplotlib==3.7.1  # Plotting library
seaborn==0.12.2  # Statistical data visualization
uvicorn==0.22.0  # ASGI server for FastAPI
loguru==0.6.0  # Logging library

# Data manipulation and analysis
dask==2023.5.0
pyarrow==12.0.1

# For working with JSON
ujson==5.7.0

# For web scraping (if needed)
beautifulsoup4==4.12.2
scrapy==2.9.0

# For handling dates and times
dateutil==2.8.2

# For managing documents and PDFs
PyMuPDF==1.20.2  # Python bindings for MuPDF
pdfminer.six==20221105  # PDF parsing library

# Environment management
conda==23.5.0

# Deployment and server management
docker==6.0.1
gunicorn==20.1.0
supervisor==4.2.4

# Testing libraries
pytest==7.3.0
pytest-cov==4.0.0
tox==4.4.0

# CI/CD and version control
pre-commit==3.3.3
black==23.3.0
isort==5.12.0

# Monitoring and performance
prometheus_client==0.16.0
opencensus-ext-prometheus==0.1.1

# Data validation
voluptuous==0.13.1
marshmallow==3.19.0

# Miscellaneous utilities
shap==0.41.0  # Explainability for ML models
pyyaml==6.0  # YAML parsing
argparse==1.4.0  # Command-line argument parsing
tabulate==0.9.0  # Pretty-print tabular data
plotly==5.14.1  # Interactive plotting

# Additional NLP libraries
gensim==4.3.0
wordcloud==1.9.2

# Image processing
Pillow==9.5.0
opencv-python==4.7.0

# Audio processing
pydub==0.25.1

# Geospatial libraries
geopy==2.3.0
shapely==2.0.1

# Networking
paramiko==3.0.0
